Fix: problems with normal runs
Set usercode.f90 back to main usercode. Fix an issue with kprocess relating to
two processes not being correctly initialised. Removed extra read line from
reader_input.f that was expecting the PDF set number. Fixed the generation of
the runscale part of the runstring so that it can correctly take the resummation
form.